Changed Types and file names for index.
