+++ tests/documents_tests.ts
@@ -330,7 +330,7 @@ describe.each([
 
     await client
       .createIndex({ uid: 'updateUid' })
-      .then((response: Types.IndexResponse) => {
+      .then((response: Types.Index) => {
         expect(response).toHaveProperty('uid', 'updateUid')
       })
     await client
